+++ crm-ui/src/app/(admin)/admin/request-management/page.tsx
@@ -6,26 +6,40 @@ import { Badge } from "@/components/ui/badge";
 import { Separator } from "@/components/ui/separator";
 import { KanbanSquare, List, Loader2, BarChart3, RefreshCw } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { KanbanView } from "@/components/kanban-view";
-import { TableView } from "@/components/table-view";
-import { TicketFilters } from "@/components/ticket-filters";
-import { Ticket, ViewMode, FilterOptions, SortOption } from "@/types/ticket";
+import { RequestKanbanView } from "@/components/request-kanban-view";
+import { RequestTableView } from "@/components/request-table-view";
+import { RequestFilters } from "@/components/request-filters";
+import { SeniorRequestDisplayView } from "@/types/request";
 import { useRequestManagement } from "@/hooks/use-requests";
 import { useToast } from "@/hooks/use-toast";
-import { TicketUtils } from "@/lib/ticket-conversion";
 
-export default function TicketManagement() {
+type ViewMode = "kanban" | "table";
+
+interface RequestFilterOptions {
+  priority?: ("low" | "medium" | "high" | "urgent")[];
+  status?: ("pending" | "in-progress" | "in-review" | "completed" | "cancelled")[];
+  requestType?: string[];
+  assignedStaff?: string[];
+  searchTerm?: string;
+}
+
+interface RequestSortOption {
+  field: 'createdAt' | 'updatedAt' | 'priority' | 'status' | 'seniorName';
+  direction: "asc" | "desc";
+}
+
+export default function RequestManagement() {
   const [mounted, setMounted] = useState(false);
   const [viewMode, setViewMode] = useState<ViewMode>("kanban");
-  const [filters, setFilters] = useState<FilterOptions>({});
-  const [sort, setSort] = useState<SortOption>({
+  const [filters, setFilters] = useState<RequestFilterOptions>({});
+  const [sort, setSort] = useState<RequestSortOption>({
     field: "createdAt",
     direction: "desc",
   });
 
   const { toast } = useToast();
 
-  // Use the new request management hook
+  // Use the request management hook
   const {
     requests,
     loading: isLoading,
@@ -42,49 +56,24 @@ export default function TicketManagement() {
     setMounted(true);
   }, []);
 
-  // Convert enhanced requests to legacy ticket format for existing components
-  const tickets: Ticket[] = useMemo(() => {
+  // Filter and sort requests based on current state
+  const processedRequests = useMemo(() => {
     if (!mounted || !requests) return [];
-    return TicketUtils.enhancedArrayToTicketArray(requests);
-  }, [mounted, requests]);
-
-  // Filter and sort tickets based on current state
-  const processedTickets = useMemo(() => {
-    if (!mounted || !tickets) return [];
     
-    // Convert FilterOptions to the new filter format
-    const newFilters = {
-      priority: filters.priority,
-      status: filters.status,
-      requestType: filters.requestType,
-      assignedStaff: filters.assignee,
-      searchTerm: filters.search,
-    };
-
-    // Convert sort field
-    const sortField = sort.field === 'seniorName' ? 'seniorName' : 
-                     sort.field === 'createdAt' ? 'createdAt' :
-                     sort.field === 'priority' ? 'priority' :
-                     sort.field === 'status' ? 'status' : 'createdAt';
-
-    // Use the enhanced requests for filtering, then convert back to tickets
-    const filteredEnhanced = filterAndSortRequests(newFilters, sortField, sort.direction);
+    // Use the enhanced requests for filtering
+    const filteredRequests = filterAndSortRequests(filters, sort.field, sort.direction);
     
-    // Convert enhanced tickets back to legacy ticket format
-    return TicketUtils.enhancedArrayToTicketArray(filteredEnhanced);
-  }, [tickets, filters, sort, viewMode, mounted, filterAndSortRequests]);
+    return filteredRequests;
+  }, [requests, filters, sort, mounted, filterAndSortRequests]);
 
-  const handleTicketUpdate = async (updatedTicket: Ticket) => {
+  const handleRequestUpdate = async (updatedRequest: SeniorRequestDisplayView) => {
     try {
-      // Convert ticket back to enhanced request format
-      const enhancedTicket = TicketUtils.ticketToEnhanced(updatedTicket);
-
-      const success = await updateRequest(enhancedTicket);
+      const success = await updateRequest(updatedRequest);
       if (!success) {
         // Error handling is done in the hook
         return;
       }
-    } catch (error) {
+    } catch {
       toast({
         title: "Error",
         description: "Failed to update senior request. Please try again.",
@@ -131,9 +120,9 @@ export default function TicketManagement() {
               Error loading requests: {error.errors[0]?.message || 'Unknown error'}
             </span>
           </div>
-          <Button onClick={refresh} className="flex items-center gap-2">
-            <RefreshCw className="h-4 w-4" />
-            Retry
+          <Button onClick={refresh} variant="outline">
+            <RefreshCw className="h-4 w-4 mr-2" />
+            Try Again
           </Button>
         </div>
       </div>
@@ -141,41 +130,35 @@ export default function TicketManagement() {
   }
 
   return (
-    <div className="min-h-screen flex flex-col bg-gray-50">
+    <div className="flex flex-col h-screen bg-gray-50">
       {/* Header */}
-      <div className="bg-white border-b border-gray-200 px-6 py-6 shadow-sm flex-shrink-0">
-        <div className="flex items-center justify-between">
-          <div>
-            <h1 className="text-3xl font-bold text-gray-900 tracking-tight">
-              Senior Care Requests
-            </h1>
-            <p className="text-gray-600 mt-2 text-lg">
-              Manage and track senior care assistance requests
-            </p>
-          </div>
+      <div className="bg-white shadow-sm border-b border-gray-200 flex-shrink-0">
+        <div className="px-6 py-4">
+          <div className="flex items-center justify-between">
+            <div>
+              <h1 className="text-2xl font-bold text-gray-900">
+                Request Management
+              </h1>
+              <p className="text-gray-600 mt-1">
+                Manage and track senior care requests
+              </p>
+            </div>
 
-          <div className="flex items-center gap-6">
+            {/* Status Summary */}
             <div className="flex items-center gap-3">
-              <Badge
-                variant="outline"
-                className="bg-slate-50 text-slate-700 border-slate-200 font-medium px-3 py-1"
-              >
-                <div className="w-2 h-2 bg-slate-400 rounded-full mr-2"></div>
-                Pending: {statusCounts.pending}
-              </Badge>
               <Badge
                 variant="outline"
                 className="bg-blue-50 text-blue-700 border-blue-200 font-medium px-3 py-1"
               >
                 <div className="w-2 h-2 bg-blue-400 rounded-full mr-2"></div>
-                In Progress: {statusCounts["in-progress"]}
+                Pending: {statusCounts.pending}
               </Badge>
               <Badge
                 variant="outline"
-                className="bg-yellow-50 text-yellow-700 border-yellow-200 font-medium px-3 py-1"
+                className="bg-orange-50 text-orange-700 border-orange-200 font-medium px-3 py-1"
               >
-                <div className="w-2 h-2 bg-yellow-400 rounded-full mr-2"></div>
-                In Review: {statusCounts["in-review"]}
+                <div className="w-2 h-2 bg-orange-400 rounded-full mr-2"></div>
+                In Progress: {statusCounts['in-progress']}
               </Badge>
               <Badge
                 variant="outline"
@@ -213,7 +196,7 @@ export default function TicketManagement() {
 
       {/* Filters */}
       <div className="flex-shrink-0">
-        <TicketFilters
+        <RequestFilters
           filters={filters}
           sort={sort}
           onFiltersChange={setFilters}
@@ -224,14 +207,14 @@ export default function TicketManagement() {
       {/* Main Content */}
       <div className="flex-1 p-6 overflow-auto min-h-0">
         {viewMode === "kanban" ? (
-          <KanbanView
-            tickets={processedTickets}
-            onTicketUpdate={handleTicketUpdate}
+          <RequestKanbanView
+            requests={processedRequests}
+            onRequestUpdate={handleRequestUpdate}
           />
         ) : (
-          <TableView
-            tickets={processedTickets}
-            onTicketUpdate={handleTicketUpdate}
+          <RequestTableView
+            requests={processedRequests}
+            onRequestUpdate={handleRequestUpdate}
           />
         )}
       </div>
@@ -247,7 +230,7 @@ export default function TicketManagement() {
             <Separator orientation="vertical" className="h-4" />
             <span className="font-medium">
               Filtered:{" "}
-              <span className="text-gray-900">{processedTickets.length}</span>
+              <span className="text-gray-900">{processedRequests.length}</span>
             </span>
           </div>
           <div className="flex items-center gap-2 text-xs text-gray-500">
@@ -258,7 +241,7 @@ export default function TicketManagement() {
               size="sm" 
               onClick={refresh}
               disabled={isLoading}
-              className="h-6 px-2 ml-2"
+              className="h-6 w-6 p-0 ml-2"
             >
               <RefreshCw className={`h-3 w-3 ${isLoading ? 'animate-spin' : ''}`} />
             </Button>
