--- crm-ui/src/components/table-view.tsx
@@ -1,392 +1,0 @@
-import { useState } from "react";
-import {
-  ColumnDef,
-  ColumnFiltersState,
-  SortingState,
-  VisibilityState,
-  flexRender,
-  getCoreRowModel,
-  getFilteredRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
-import { Avatar, AvatarFallback } from "@/components/ui/avatar";
-import {
-  ChevronLeft,
-  ChevronRight,
-  ChevronsLeft,
-  ChevronsRight,
-  ArrowUpDown,
-  AlertTriangle,
-} from "lucide-react";
-import { Ticket } from "@/types/ticket";
-import {
-  getPriorityColor,
-  getStatusColor,
-  formatDate,
-  isOverdue,
-} from "@/lib/ticket-utils";
-import { TicketModal } from "@/components/ticket-modal";
-import { cn } from "@/lib/utils";
-
-interface TableViewProps {
-  tickets: Ticket[];
-  onTicketUpdate: (ticket: Ticket) => void;
-}
-
-export function TableView({ tickets, onTicketUpdate }: TableViewProps) {
-  const [sorting, setSorting] = useState<SortingState>([]);
-  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
-  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
-  const [selectedTicket, setSelectedTicket] = useState<Ticket | null>(null);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const getInitials = (name: string) => {
-    return name
-      .split(" ")
-      .map((n) => n[0])
-      .join("")
-      .toUpperCase();
-  };
-
-  const columns: ColumnDef<Ticket>[] = [
-    {
-      accessorKey: "id",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Ticket ID
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => (
-        <Badge
-          variant="secondary"
-          className="font-mono text-xs bg-gray-50 text-gray-700 hover:bg-gray-50"
-        >
-          {row.getValue("id")}
-        </Badge>
-      ),
-    },
-    {
-      accessorKey: "title",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Title
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => (
-        <div className="font-medium max-w-[300px] truncate text-gray-900">
-          {row.getValue("title")}
-        </div>
-      ),
-    },
-    {
-      accessorKey: "priority",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Priority
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => {
-        const priority = row.getValue("priority") as string;
-        return (
-          <Badge
-            variant="outline"
-            className={cn(
-              "text-xs font-medium",
-              getPriorityColor(priority as Ticket["priority"])
-            )}
-          >
-            {priority.charAt(0).toUpperCase() + priority.slice(1)}
-          </Badge>
-        );
-      },
-    },
-    {
-      accessorKey: "status",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Status
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => {
-        const status = row.getValue("status") as string;
-        return (
-          <Badge
-            variant="outline"
-            className={cn(
-              "text-xs font-medium",
-              getStatusColor(status as Ticket["status"])
-            )}
-          >
-            {status.replace("-", " ").replace(/\b\w/g, (l) => l.toUpperCase())}
-          </Badge>
-        );
-      },
-    },
-    {
-      accessorKey: "assignee",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Assignee
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => {
-        const assignee = row.getValue("assignee") as string;
-        return (
-          <div className="flex items-center gap-2">
-            <Avatar className="h-6 w-6">
-              <AvatarFallback className="text-xs bg-gray-100 text-gray-600 font-medium">
-                {getInitials(assignee)}
-              </AvatarFallback>
-            </Avatar>
-            <span className="text-sm font-medium text-gray-900">
-              {assignee}
-            </span>
-          </div>
-        );
-      },
-    },
-    {
-      accessorKey: "dueDate",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Due Date
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => {
-        const dueDate = row.getValue("dueDate") as Date;
-        const ticket = row.original;
-        const overdue = isOverdue(dueDate) && ticket.status !== "completed";
-
-        return (
-          <div
-            className={cn(
-              "flex items-center gap-1 text-sm font-medium",
-              overdue ? "text-red-600" : "text-gray-700"
-            )}
-          >
-            {overdue && <AlertTriangle className="h-4 w-4" />}
-            <span>{formatDate(dueDate)}</span>
-          </div>
-        );
-      },
-    },
-    {
-      accessorKey: "createdDate",
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-          className="h-8 p-0 hover:bg-transparent font-semibold text-gray-900"
-        >
-          Created
-          <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => (
-        <div className="text-sm text-gray-600 font-medium">
-          {formatDate(row.getValue("createdDate"))}
-        </div>
-      ),
-    },
-  ];
-
-  const table = useReactTable({
-    data: tickets,
-    columns,
-    onSortingChange: setSorting,
-    onColumnFiltersChange: setColumnFilters,
-    getCoreRowModel: getCoreRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    onColumnVisibilityChange: setColumnVisibility,
-    state: {
-      sorting,
-      columnFilters,
-      columnVisibility,
-    },
-    initialState: {
-      pagination: {
-        pageSize: 10,
-      },
-    },
-  });
-
-  return (
-    <div className="space-y-4">
-      <div className="rounded-lg border border-gray-200 bg-white shadow-sm">
-        <Table>
-          <TableHeader>
-            {table.getHeaderGroups().map((headerGroup) => (
-              <TableRow
-                key={headerGroup.id}
-                className="hover:bg-transparent border-b border-gray-200"
-              >
-                {headerGroup.headers.map((header) => (
-                  <TableHead
-                    key={header.id}
-                    className="font-semibold text-gray-900 bg-gray-50"
-                  >
-                    {header.isPlaceholder
-                      ? null
-                      : flexRender(
-                          header.column.columnDef.header,
-                          header.getContext()
-                        )}
-                  </TableHead>
-                ))}
-              </TableRow>
-            ))}
-          </TableHeader>
-          <TableBody>
-            {table.getRowModel().rows?.length ? (
-              table.getRowModel().rows.map((row) => (
-                <TableRow
-                  key={row.id}
-                  data-state={row.getIsSelected() && "selected"}
-                  className="hover:bg-gray-50 cursor-pointer border-b border-gray-100 transition-colors"
-                  onClick={() => {
-                    setSelectedTicket(row.original);
-                    setIsModalOpen(true);
-                  }}
-                >
-                  {row.getVisibleCells().map((cell) => (
-                    <TableCell key={cell.id} className="py-4">
-                      {flexRender(
-                        cell.column.columnDef.cell,
-                        cell.getContext()
-                      )}
-                    </TableCell>
-                  ))}
-                </TableRow>
-              ))
-            ) : (
-              <TableRow>
-                <TableCell
-                  colSpan={columns.length}
-                  className="h-24 text-center text-gray-500"
-                >
-                  No tickets found.
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-      </div>
-
-      {/* Pagination */}
-      <div className="flex items-center justify-between bg-white px-4 py-3 border border-gray-200 rounded-lg">
-        <div className="text-sm text-gray-600 font-medium">
-          Showing{" "}
-          {table.getState().pagination.pageIndex *
-            table.getState().pagination.pageSize +
-            1}{" "}
-          to{" "}
-          {Math.min(
-            (table.getState().pagination.pageIndex + 1) *
-              table.getState().pagination.pageSize,
-            table.getFilteredRowModel().rows.length
-          )}{" "}
-          of {table.getFilteredRowModel().rows.length} tickets
-        </div>
-
-        <div className="flex items-center space-x-2">
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.setPageIndex(0)}
-            disabled={!table.getCanPreviousPage()}
-            className="border-gray-300"
-          >
-            <ChevronsLeft className="h-4 w-4" />
-          </Button>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.previousPage()}
-            disabled={!table.getCanPreviousPage()}
-            className="border-gray-300"
-          >
-            <ChevronLeft className="h-4 w-4" />
-          </Button>
-          <span className="text-sm font-medium text-gray-700 px-2">
-            Page {table.getState().pagination.pageIndex + 1} of{" "}
-            {table.getPageCount()}
-          </span>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.nextPage()}
-            disabled={!table.getCanNextPage()}
-            className="border-gray-300"
-          >
-            <ChevronRight className="h-4 w-4" />
-          </Button>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.setPageIndex(table.getPageCount() - 1)}
-            disabled={!table.getCanNextPage()}
-            className="border-gray-300"
-          >
-            <ChevronsRight className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
-
-      {selectedTicket && (
-        <TicketModal
-          ticket={selectedTicket}
-          isOpen={isModalOpen}
-          onOpenChange={setIsModalOpen}
-          onUpdate={(updatedTicket) => {
-            onTicketUpdate(updatedTicket);
-            setIsModalOpen(false);
-          }}
-        />
-      )}
-    </div>
-  );
-}
