+++ crm-ui/src/components/ui/droppable-column.tsx
@@ -1,0 +1,31 @@
+import { useDroppable } from "@dnd-kit/core";
+import { cn } from "@/lib/utils";
+
+interface DroppableColumnProps {
+  id: string;
+  children: React.ReactNode;
+  className?: string;
+}
+
+export function DroppableColumn({
+  id,
+  children,
+  className,
+}: DroppableColumnProps) {
+  const { setNodeRef, isOver } = useDroppable({
+    id,
+  });
+
+  return (
+    <div
+      ref={setNodeRef}
+      className={cn(
+        "min-h-[200px] rounded-lg p-2 transition-all duration-200",
+        isOver && "bg-blue-50 border-2 border-blue-300 border-dashed",
+        className
+      )}
+    >
+      {children}
+    </div>
+  );
+}
