+++ crm-ui/src/types/request.ts
@@ -43,7 +43,7 @@ export interface SeniorRequestFilterDto {
   createdBefore?: string; // ISO string
 }
 
-// High-performance projection for read operations
+// High-performance projection for read operations - matches backend SeniorRequestView interface
 export interface SeniorRequestView {
   id: number;
   seniorId: number;
@@ -56,6 +56,25 @@ export interface SeniorRequestView {
   updatedAt: string;
   completedAt?: string;
   status: RequestStatus;
+}
+
+// Extended view with populated reference data for UI display
+export interface SeniorRequestDisplayView extends SeniorRequestView {
+  // Additional senior information (populated via join or separate call)
+  seniorName?: string;
+  seniorPhone?: string;
+  seniorEmail?: string;
+  seniorAddress?: string;
+  
+  // Additional staff information (populated via join or separate call)
+  assignedStaffName?: string;
+  
+  // Additional request type information
+  requestTypeName?: string;
+  
+  // Frontend compatibility fields (computed from backend data)
+  frontendStatus: 'pending' | 'in-progress' | 'completed' ;
+  frontendPriority: 'low' | 'medium' | 'high' | 'urgent';
 }
 
 // Request types from backend
@@ -77,42 +96,10 @@ export interface StaffDto {
   updatedAt: string;
 }
 
-// Enhanced ticket type that bridges frontend and backend
-export interface EnhancedTicket {
-  // Backend fields
-  id: number;
-  seniorId: number;
-  assignedStaffId?: number;
-  requestTypeId?: number;
-  title: string;
-  description: string;
-  priority: number;
-  status: RequestStatus;
-  createdAt: string;
-  updatedAt: string;
-  completedAt?: string;
-  
-  // Additional senior information (populated via join or separate call)
-  seniorName?: string;
-  seniorPhone?: string;
-  seniorEmail?: string;
-  seniorAddress?: string;
-  
-  // Additional staff information (populated via join or separate call)
-  assignedStaffName?: string;
-  
-  // Additional request type information
-  requestTypeName?: string;
-  
-  // Frontend compatibility fields (computed from backend data)
-  frontendStatus: 'pending' | 'in-progress' | 'in-review' | 'completed' | 'cancelled';
-  frontendPriority: 'low' | 'medium' | 'high' | 'urgent';
-}
-
 // Utility functions for converting between backend and frontend formats
 export const RequestUtils = {
   // Convert backend status to frontend status
-  backendToFrontendStatus(status: RequestStatus): 'pending' | 'in-progress' | 'in-review' | 'completed' | 'cancelled' {
+  backendToFrontendStatus(status: RequestStatus): 'pending' | 'in-progress'  | 'completed'  {
     switch (status) {
       case 'TODO':
         return 'pending';
@@ -126,17 +113,14 @@ export const RequestUtils = {
   },
 
   // Convert frontend status to backend status
-  frontendToBackendStatus(status: 'pending' | 'in-progress' | 'in-review' | 'completed' | 'cancelled'): RequestStatus {
+  frontendToBackendStatus(status: 'pending' | 'in-progress'  | 'completed' ): RequestStatus {
     switch (status) {
       case 'pending':
         return 'TODO';
       case 'in-progress':
-      case 'in-review': // Map both to IN_PROGRESS since backend only has 3 states
         return 'IN_PROGRESS';
       case 'completed':
         return 'COMPLETED';
-      case 'cancelled':
-        return 'TODO'; // Could be enhanced to add CANCELLED status to backend
       default:
         return 'TODO';
     }
@@ -166,15 +150,32 @@ export const RequestUtils = {
     }
   },
 
-  // Convert SeniorRequestDto to EnhancedTicket
-  toEnhancedTicket(request: SeniorRequestDto, additionalInfo?: {
+  // Convert SeniorRequestView to SeniorRequestDisplayView with additional info
+  toDisplayView(request: SeniorRequestView, additionalInfo?: {
     seniorName?: string;
     seniorPhone?: string;
     seniorEmail?: string;
     seniorAddress?: string;
     assignedStaffName?: string;
     requestTypeName?: string;
-  }): EnhancedTicket {
+  }): SeniorRequestDisplayView {
+    return {
+      ...request,
+      ...additionalInfo,
+      frontendStatus: this.backendToFrontendStatus(request.status),
+      frontendPriority: this.backendToFrontendPriority(request.priority),
+    };
+  },
+
+  // Convert SeniorRequestDto to SeniorRequestDisplayView with additional info
+  fromDtoToDisplayView(request: SeniorRequestDto, additionalInfo?: {
+    seniorName?: string;
+    seniorPhone?: string;
+    seniorEmail?: string;
+    seniorAddress?: string;
+    assignedStaffName?: string;
+    requestTypeName?: string;
+  }): SeniorRequestDisplayView {
     return {
       ...request,
       ...additionalInfo,
